LPD-90910 Skip commerce category friendly URL handling under LPD-70396

+++ modules/apps/commerce/commerce-product-service/src/main/java/com/liferay/commerce/product/internal/model/listener/AssetCategoryModelListener.java
@@ -11,6 +11,7 @@ import com.liferay.commerce.product.service.CPDisplayLayoutLocalService;
 import com.liferay.friendly.url.service.FriendlyURLEntryLocalService;
 import com.liferay.portal.kernel.exception.ModelListenerException;
 import com.liferay.portal.kernel.exception.PortalException;
+import com.liferay.portal.kernel.feature.flag.FeatureFlagManagerUtil;
 import com.liferay.portal.kernel.log.Log;
 import com.liferay.portal.kernel.log.LogFactoryUtil;
 import com.liferay.portal.kernel.model.BaseModelListener;
@@ -37,6 +38,12 @@ public class AssetCategoryModelListener
 	public void onAfterCreate(AssetCategory assetCategory)
 		throws ModelListenerException {
 
+		if (FeatureFlagManagerUtil.isEnabled(
+				assetCategory.getCompanyId(), "LPD-70396")) {
+
+			return;
+		}
+
 		try {
 			_friendlyURLEntryLocalService.addFriendlyURLEntry(
 				assetCategory.getGroupId(),
@@ -59,6 +66,12 @@ public class AssetCategoryModelListener
 
 			_cpDisplayLayoutLocalService.deleteCPDisplayLayouts(
 				AssetCategory.class, assetCategory.getCategoryId());
+
+			if (FeatureFlagManagerUtil.isEnabled(
+					assetCategory.getCompanyId(), "LPD-70396")) {
+
+				return;
+			}
 
 			_friendlyURLEntryLocalService.deleteFriendlyURLEntry(
 				assetCategory.getGroupId(),
